Carefully move demangle.h from meta to utils
