+++ stub/include/neurala/utils/detail/AnyRef.h
@@ -24,7 +24,7 @@
 #include <typeindex>
 
 #include "neurala/exports.h"
-#include "neurala/meta/detail/demangle.h"
+#include "neurala/utils/demangle.h"
 #include "neurala/meta/detail/typeName.h"
 #include "neurala/utils/string.h"
